PhysicalFileSystem.DeleteDirectory is a single line wrapper around for PhysicalFileSystem.RecursiveDelete. Removing redundant DeleteDirectory.

Fixes #758

+++ GVFS/GVFS.Common/FileSystem/PhysicalFileSystem.cs
@@ -12,7 +12,7 @@ namespace GVFS.Common.FileSystem
     {
         public const int DefaultStreamBufferSize = 8192;
 
-        public virtual void RecursiveDelete(string path)
+        public virtual void DeleteDirectory(string path, bool recursive = true)
         {
             if (!Directory.Exists(path))
             {
@@ -21,15 +21,18 @@ namespace GVFS.Common.FileSystem
 
             DirectoryInfo directory = new DirectoryInfo(path);
 
-            foreach (FileInfo file in directory.GetFiles())
+            if (recursive)
             {
-                file.Attributes = FileAttributes.Normal;
-                file.Delete();
-            }
+                foreach (FileInfo file in directory.GetFiles())
+                {
+                    file.Attributes = FileAttributes.Normal;
+                    file.Delete();
+                }
 
-            foreach (DirectoryInfo subDirectory in directory.GetDirectories())
-            {
-                this.RecursiveDelete(subDirectory.FullName);
+                foreach (DirectoryInfo subDirectory in directory.GetDirectories())
+                {
+                    this.DeleteDirectory(subDirectory.FullName);
+                }
             }
 
             directory.Delete();
@@ -124,11 +127,6 @@ namespace GVFS.Common.FileSystem
         public virtual bool TryCreateDirectoryWithAdminOnlyModify(ITracer tracer, string directoryPath, out string error)
         {
             return GVFSPlatform.Instance.FileSystem.TryCreateDirectoryWithAdminOnlyModify(tracer, directoryPath, out error);
-        }
-
-        public virtual void DeleteDirectory(string path, bool recursive = false)
-        {
-            this.RecursiveDelete(path);
         }
 
         public virtual bool IsSymLink(string path)
